backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "open-deep-research-web"
version = "0.1.0"
description = "Web implementation of Open Deep Research Agent with Ollama"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
langchain-ollama = "^0.2.0"
pydantic = "^2.8.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
tavily-python = "^0.5.0"
websockets = "^13.0"
aiohttp = "^3.10.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
alembic = "^1.13.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2.0"
langdetect = "^1.0.9"
python-dotenv = "^1.0.1"
python-socketio = "^5.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
black = "^24.8.0"
mypy = "^1.11.0"
ruff = "^0.6.0"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["E501", "W503"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true